requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
pandas==2.1.4
openpyxl==3.1.2
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0 
django-cors-headers==4.3.1